+++ packages/orq-rc/src/orq_ai_sdk/models/createcontactop.py
@@ -134,7 +134,7 @@ class CreateContactResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-04-22T12:00:25.416Z")
+    updated: Optional[datetime] = parse_datetime("2025-04-23T11:10:24.059Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")