+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JSENY2W8PMG2T1NFE18PNYDE"
+        "01JSH5F6WDSZDHVDVPAM1QSWTE"
     )
     r"""The id of the resource"""