+++ packages/orq-rc/src/orq_ai_sdk/models/createtoolop.py
@@ -448,7 +448,7 @@ class ResponseBody3(BaseModel):
     http: ResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01JSENY2TZ5FYS3KQHA5R6K737"
+        "tool_01JSH5F6S38D9ZWS6C1KDMS391"
     )
 
     created_by_id: Optional[str] = None
@@ -570,7 +570,7 @@ class ResponseBody2(BaseModel):
     json_schema: ResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01JSENY2TYVY7W4GQ03NVWVVN2"
+        "tool_01JSH5F6S3V4YMRX0WNS5GVXAQ"
     )
 
     created_by_id: Optional[str] = None
@@ -662,7 +662,7 @@ class ResponseBody1(BaseModel):
     function: ResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01JSENY2TX9NQ1NCXNJCZP25S6"
+        "tool_01JSH5F6RZ366YZPXFM49DRH3B"
     )
 
     created_by_id: Optional[str] = None